# [VLM] Improving multimodal tensor hash kernel