## Motivation
Improving hash kernel security+performance wise 
https://github.com/sgl-project/sglang/issues/8054
https://github.com/sgl-project/sglang/issues/8342

current hash kernel uses `.sum()` - bad reduction and no avalanche properties.

## Modifications

The new hash function. we do a per-word mixing using Murmur3 fmix32.
Final 64-bit SplitMix64 runs on CPU.

## Benchmark & Profiling
Resulting hashing mechanism is safer and is as fast the previous method.

| Model         | Old Time (s) | Old Acc  | New Time (s) | New Acc  |
|---------------|--------------|----------|--------------|----------|
| Qwen2 7B VL   | 372.1437     | 0.498    | 369.1880     | 0.496    |
| Gemma 3 4B it | 55.8040      | 0.348    | 55.1853      | 0.342    |
| MiniCPMV 2_6  | 270.2827     | 0.449    | 268.7281     | 0.453    |

Benchmarking across various tensor sizes:

| Size (elements) | Old Avg Time (s) | Old Throughput (GB/s) | New Avg Time (s) | New Throughput (GB/s) |
|-----------------|------------------|-----------------------|------------------|-----------------------|
| 1,024           | 0.000151         | 0.03                  | 0.000131         | 0.03                  |
| 16,384          | 0.000175         | 0.37                  | 0.000176         | 0.37                  |
| 1,048,576       | 0.000228         | 18.41                  | 0.000195         | 21.48                 |


## Checklist

- [X] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).