@mickqian 2025-08-14
Well done Adarsh!
Also, be aware that for the previous hash kernel, some bugs would take place, relating to TP, or non-fast-image-processor.
We should pay constant attention to this change, since this the fundamental for VLMs