@adarshxs 2025-08-14
Hey Mick,

Here are the results from the collision test script:
**Previous Hash:**
```
Tested 5,000 random tensors of shape (256,)
Collisions found: 0
Collision rate: 0.000000
Total time: 3.071572 sec
Average time per tensor: 614.31 µs
```

**New Hash:**
```
Hasher: GPU Tensor Hash
Tested 5,000 random tensors of shape (256,)
Collisions found: 0
Collision rate: 0.000000
Total time: 2.166767 sec
Average time per tensor: 433.35 µs
```

I'll keep in mind the bugs with TP and non fast image processor. Any suggestions/tests you would like me to run?